+++ NEWS.md
@@ -2,6 +2,7 @@
 - Improve backward compatibility of R function `compute_analogs`. This function now returns the analog member index in the 4D observations, rather than in the 3D observations.
 - Add a cleaner version of configuration. Configuration now, by default, is the cleaner version. The old version can be accessed by using `advanced` parameter when generating a configuration.
 - Remove some of the post contents to reduce redundant information. Tutorials are now only hosted on binder.
+- Add MPI implementation for "MPI_UNIVERSE_SIZE" to limit the number of child spawned processes.
 
 # PEF 3.2.1